# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Full Express API",
    "version": "1.0.0",
    "description": "A comprehensive Express API with Prisma"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://back-end-production-17b6.up.railway.app/api",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/quizzes": {
      "get": {
        "summary": "Get all quizzes",
        "tags": [
          "Quizzes"
        ],
        "responses": {
          "200": {
            "description": "List of all quizzes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quiz"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new quiz",
        "tags": [
          "Quizzes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quiz created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/quizzes/{id}": {
      "get": {
        "summary": "Get a quiz by ID",
        "tags": [
          "Quizzes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Quiz ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found"
          }
        }
      },
      "put": {
        "summary": "Update a quiz",
        "tags": [
          "Quizzes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Quiz ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quiz updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a quiz",
        "tags": [
          "Quizzes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Quiz ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Quiz deleted successfully"
          },
          "404": {
            "description": "Quiz not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Quiz": {
        "type": "object",
        "required": [
          "title",
          "moduleId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the quiz"
          },
          "title": {
            "type": "string",
            "description": "Quiz title"
          },
          "description": {
            "type": "string",
            "description": "Quiz description"
          },
          "moduleId": {
            "type": "string",
            "description": "ID of the module this quiz belongs to"
          },
          "timeLimit": {
            "type": "integer",
            "description": "Time limit in minutes"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the quiz is active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Quiz creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Quiz last update timestamp"
          }
        }
      },
      "QuizInput": {
        "type": "object",
        "required": [
          "title",
          "moduleId"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "moduleId": {
            "type": "string"
          },
          "timeLimit": {
            "type": "integer"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}